compile init.js
