+++ .changes/fix-invoke-usage.md
@@ -1,0 +1,6 @@
+---
+"shell": patch
+"dialog": patch
+---
+
+Fix invoke usage.
