+++ plugins/shell/src/lib.rs
@@ -78,7 +78,7 @@ impl<R: Runtime, T: Manager<R>> ShellExt<R> for T {
 }
 
 pub fn init<R: Runtime>() -> TauriPlugin<R, Option<Config>> {
-    let mut init_script = include_str!("init.js").to_string();
+    let mut init_script = include_str!("init-iife.js").to_string();
     init_script.push_str(include_str!("api-iife.js"));
 
     Builder::<R, Option<Config>>::new("shell")
